refactor(products): implement useProducts to ProductCollection #103
