--- client/src/components/ProductCollection/ProductCollection.jsx
@@ -1,62 +1,0 @@
-import { useState } from "react";
-import { useEffect } from "react";
-import ProductCarousel from "../ProductCarousel/ProductCarousel";
-import { SERVER_URL } from "../../config.js";
-
-export default function ProductCollection({
-  category,
-  sortOption,
-  isFeatured,
-  title,
-  subcategory,
-  limit,
-}) {
-  const [products, setProducts] = useState(null);
-
-  useEffect(() => {
-    const getProducts = async () => {
-      let url = `${SERVER_URL}/api/v1/products`;
-
-      let params = "";
-
-      if (sortOption) {
-        params += `&sort=${sortOption}`;
-      }
-      if (category) {
-        params += `&category=${category}`;
-      }
-      if (isFeatured) {
-        params += `&isFeatured=${isFeatured}`;
-      }
-      if (subcategory) {
-        params += `&subcategory=${subcategory}`;
-      }
-      if (limit) {
-        params += `&limit=${limit}`;
-      }
-
-      if (params) url += "?" + params.substring(1); // removes the 1st char of params not(0) it should be (1)
-
-      try {
-        const res = await fetch(url);
-        const data = await res.json();
-        setProducts(data.products);
-      } catch (error) {
-        console.error(`Error fetching ${url} `, error);
-      }
-    };
-    getProducts();
-  }, [category, sortOption, isFeatured, subcategory, limit]); // slug and title is excluded so the comp doesnt re-render
-
-  if (!products) return <p>Loading...</p>;
-
-  if (products.length === 0) return <p>No product Found...</p>;
-
-  return (
-    <ProductCarousel
-      products={products}
-      title={title}
-      viewAllLink={`../products?subcategory=${subcategory}`}
-    />
-  );
-}
